Usage comment for script
This script is only used for "calibration", or when a different Arduino is used. In the RPi server file, the serial number of the Arduino is hardcoded so that the RPi can connect to the serial port no matter which usb port it is plugged into. If a different Arduino is used, this number will change and the RPi server file needs to be updated to the new serial number.