These are old files from the first version of avionics
SOPs describe how to perform an action. main.py and RocketController.py should have examples on how to implement bluetooth on the RPi and how to implement usb serial with an arduino. usbScan.py will give information about devices connected over usb to the RPi, you will need it for a parameter in RocketController.py Arduino_Code is the code for the arduino connected to the Pi. bluetooth_client.py is run on a computer to connected to the Pi bluetooth connection.